# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/auth.UserDTO'
    type: object
  auth.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  auth.UserDTO:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  bracket.BracketNode:
    properties:
      game_id:
        type: string
      id:
        type: string
      left_child:
        $ref: '#/definitions/bracket.BracketNode'
      position:
        type: integer
      right_child:
        $ref: '#/definitions/bracket.BracketNode'
      round:
        type: integer
      scheduled_time:
        type: string
      status:
        type: string
      team1_id:
        type: string
      team1_name:
        type: string
      team1_score:
        type: integer
      team1_seed:
        type: integer
      team2_id:
        type: string
      team2_name:
        type: string
      team2_score:
        type: integer
      team2_seed:
        type: integer
      winner_id:
        type: string
    type: object
  bracket.BracketType:
    enum:
    - single_elimination
    - double_elimination
    type: string
    x-enum-varnames:
    - BracketTypeSingleElimination
    - BracketTypeDoubleElimination
  bracket.GenerateBracketRequest:
    properties:
      bracket_type:
        allOf:
        - $ref: '#/definitions/bracket.BracketType'
        enum:
        - single_elimination
        - double_elimination
      event_id:
        type: string
      field_id:
        type: string
      game_duration:
        minimum: 30
        type: integer
      round_id:
        type: string
      start_time:
        type: string
      teams:
        items:
          $ref: '#/definitions/bracket.TeamSeed'
        minItems: 2
        type: array
    required:
    - bracket_type
    - event_id
    - field_id
    - game_duration
    - round_id
    - start_time
    - teams
    type: object
  bracket.GenerateBracketResponse:
    properties:
      bracket_id:
        type: string
      bracket_tree:
        $ref: '#/definitions/bracket.BracketNode'
      bracket_type:
        $ref: '#/definitions/bracket.BracketType'
      created_at:
        type: string
      event_id:
        type: string
      games_created:
        items:
          type: string
        type: array
      round_id:
        type: string
      total_games:
        type: integer
      total_rounds:
        type: integer
    type: object
  bracket.GetBracketResponse:
    properties:
      bracket_tree:
        $ref: '#/definitions/bracket.BracketNode'
      bracket_type:
        $ref: '#/definitions/bracket.BracketType'
      event_id:
        type: string
      round_id:
        type: string
      total_games:
        type: integer
      total_rounds:
        type: integer
      updated_at:
        type: string
    type: object
  bracket.TeamSeed:
    properties:
      seed:
        minimum: 1
        type: integer
      team_id:
        type: string
      team_name:
        type: string
    required:
    - seed
    - team_id
    - team_name
    type: object
...
  handlers.ErrorResponse:
    description: Error response structure
    properties:
      error:
        example: error message
        type: string
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for DigiGameStats reimplementation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DigiGameStats API
  version: "1.0"
paths:
...
  /api/v1/events/{id}/bracket:
    get:
      description: Get the bracket structure for an event's bracket round
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bracket.GetBracketResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get event bracket
      tags:
      - brackets
  /api/v1/events/{id}/generate-bracket:
    post:
      consumes:
      - application/json
      description: Generate a tournament bracket structure with games
      parameters:
      - description: Bracket generation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/bracket.GenerateBracketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/bracket.GenerateBracketResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Generate tournament bracket
      tags:
      - brackets
  /api/v1/rounds/{id}/bracket:
    get:
      description: Get the bracket structure for a specific round
      parameters:
      - description: Round ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bracket.GetBracketResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get round bracket
      tags:
      - brackets
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return access/refresh tokens.
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "401":
          description: unauthorized
          schema:
            type: string
      summary: Login
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token using refresh token.
      parameters:
      - description: Refresh Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "401":
          description: unauthorized
          schema:
            type: string
      summary: Refresh Token
      tags:
      - auth
...
  /me:
    get:
      description: Get information about the currently authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.UserDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Current User
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"